Fix errors in 0036_convert_address_textfield and sample data
The migration didn't deal with the case when there was a clean database, I don't believe
a ContentType would exist in that situation, which caused the migration to break. This
fixes that by not assuming one exists already.

The sample data also should only contain swoptact data rather than a datadump
from all the apps.

+++ swoptact/migrations/0036_convert_address_textfield.py
@@ -71,11 +71,13 @@ def remove_address_contenttype(apps, schema_editor):
         be confused and avoidable.
     """
     ContentType = apps.get_model("contenttypes", "ContentType")
-    address_ct = ContentType.objects.get(
+    address_ct = ContentType.objects.filter(
         app_label="swoptact",
         model="address"
     )
-    address_ct.delete()
+
+    if address_ct.exists():
+        address_ct.delete()
 
 class Migration(migrations.Migration):
